Make condor fully compatible with Python 3

+++ examples/scripts/rotations/example.py
@@ -46,7 +46,7 @@ for angle_d in angles_d:
     #rotation_formalism = None
     #rotation_mode = "extrinsic"
 
-    #print "Angle = %.2f degrees" % angle_d
+    #print("Angle = %.2f degrees" % angle_d)
 
     short_diameter = 25E-9*12/100.
     long_diameter = 2*short_diameter
@@ -58,7 +58,7 @@ for angle_d in angles_d:
     # Spheroid
     if True:
         # Ideal spheroid
-        #print "Simulating spheroid"
+        #print("Simulating spheroid")
         par = condor.ParticleSpheroid(diameter=spheroid_diameter, material_type="water", flattening=spheroid_flattening, rotation_values=rotation_values, rotation_formalism=rotation_formalism, rotation_mode=rotation_mode)
         s = "particle_spheroid"
         E = condor.Experiment(src, {s : par}, det)
@@ -71,7 +71,7 @@ for angle_d in angles_d:
     
     if True:
         # Map (spheroid)
-        #print "Simulating map (spheroid)"
+        #print("Simulating map (spheroid)")
         par = condor.ParticleMap(diameter=spheroid_diameter, material_type="water", flattening=spheroid_flattening, geometry="spheroid", rotation_values=rotation_values, rotation_formalism=rotation_formalism, rotation_mode=rotation_mode)
         s = "particle_map_spheroid"
         E = condor.Experiment(src, {s : par}, det)
@@ -90,7 +90,7 @@ for angle_d in angles_d:
         map3d[:N_short,:,:N_short] = 1.
         map3d[N_short:N_short+N_short,:N_short,:N_short] = 1.
         # Map
-        #print "Simulating map (custom)"
+        #print("Simulating map (custom)")
         par = condor.ParticleMap(diameter=long_diameter, material_type="water", geometry="custom", map3d=map3d, dx=dx, rotation_values=rotation_values, rotation_formalism=rotation_formalism, rotation_mode=rotation_mode)
         s = "particle_map_custom"
         E = condor.Experiment(src, {s : par}, det)
@@ -107,7 +107,7 @@ for angle_d in angles_d:
 
     if True:
         # Atoms (box)
-        #print "Simulating atoms"
+        #print("Simulating atoms")
         Z1,Y1,X1 = numpy.meshgrid(numpy.linspace(0, short_diameter, N_short),
                                   numpy.linspace(0, long_diameter,   N_long),
                                   numpy.linspace(0, short_diameter, N_short),
@@ -126,7 +126,7 @@ for angle_d in angles_d:
         if plotting:
             pypl.imsave(out_dir + "/%s_proj.png" % (s),proj)
         atomic_positions = numpy.array([[x,y,z] for x,y,z in zip(X.ravel(),Y.ravel(),Z.ravel())])
-        atomic_numbers   = numpy.ones(atomic_positions.size/3, dtype=numpy.int16)
+        atomic_numbers   = numpy.ones(int(atomic_positions.size/3), dtype=numpy.int16)
         par = condor.ParticleAtoms(atomic_positions=atomic_positions, atomic_numbers=atomic_numbers, rotation_values=rotation_values, rotation_formalism=rotation_formalism, rotation_mode=rotation_mode)
         s = "particle_atoms"
         E = condor.Experiment(src, {s : par}, det)
